Moved inventory display into item instances.
Seeing how in the future each individual item will get more and more unique as the game goes on, I had the player call a script, which in turn called a user defined event in the specific item to draw, instead of just having `obj_player` draw it. This will allow each item to handle the drawing individually.