Created Issue Templates
**PLEASE READ**

I thought Isaiah's idea for a template to follow when creating issues was really great and, like with most things, I wondered if it didn't already exist. To my delight, I found a very integrated issue template system. I put in Isaiah's templates and I am commiting them now! You should now be prompted to choose a template when you create an issue.

Feel free to leave the headers as you delete and fill in the prompts, or maybe we don't need the headers.  If you think you think you can improve one of these templates, or want to add a new one, head to the repository "Settings", and under the "Features" header on the "Options" page you will find a big green button saying "Set up templates." I only came up with a Bug report, Feature request, and one for all other General issues so there may be more to add!

Cheers!